further cleanup, addressing linter issues
