+++ .golangci.yml
@@ -61,10 +61,10 @@ linters-settings:
   misspell:
     locale: US
   gocyclo:
-    min-complexity: 10
+    min-complexity: 14
   funlen:
     lines: 120
-    statements: 40
+    statements: 60
 
 issues:
   max-issues-per-linter: 0
