+++ indicators/bbands.go
@@ -2,7 +2,7 @@ package indicators
 
 // BBands - Bollinger Bands
 // upperband, middleband, lowerband = BBands(close, timeperiod=5, nbdevup=2, nbdevdn=2, matype=0)
-func BBands(inReal []float64, inTimePeriod int, inNbDevUp float64, inNbDevDn float64, inMAType MaType) ([]float64, []float64, []float64) {
+func BBands(inReal []float64, inTimePeriod int, inNbDevUp, inNbDevDn float64, inMAType MaType) ([]float64, []float64, []float64) {
 	outRealUpperBand := make([]float64, len(inReal))
 	outRealMiddleBand := Ma(inReal, inTimePeriod, inMAType)
 	outRealLowerBand := make([]float64, len(inReal))
