+++ indicators/oscillator.go
@@ -1,7 +1,9 @@
 package indicators
 
+import "math"
+
 // Apo - Absolute Price Oscillator
-func Apo(inReal []float64, inFastPeriod int, inSlowPeriod int, inMAType MaType) []float64 {
+func Apo(inReal []float64, inFastPeriod, inSlowPeriod int, inMAType MaType) []float64 {
 	if inSlowPeriod < inFastPeriod {
 		inSlowPeriod, inFastPeriod = inFastPeriod, inSlowPeriod
 	}
@@ -94,7 +96,7 @@ func Cmo(inReal []float64, inTimePeriod int) []float64 {
 }
 
 // Ppo - Percentage Price Oscillator
-func Ppo(inReal []float64, inFastPeriod int, inSlowPeriod int, inMAType MaType) []float64 {
+func Ppo(inReal []float64, inFastPeriod, inSlowPeriod int, inMAType MaType) []float64 {
 	if inSlowPeriod < inFastPeriod {
 		inSlowPeriod, inFastPeriod = inFastPeriod, inSlowPeriod
 	}
@@ -114,7 +116,7 @@ func Ppo(inReal []float64, inFastPeriod int, inSlowPeriod int, inMAType MaType) 
 }
 
 // AdOsc - Chaikin A/D Oscillator
-func AdOsc(inHigh []float64, inLow []float64, inClose []float64, inVolume []float64, inFastPeriod int, inSlowPeriod int) []float64 {
+func AdOsc(inHigh, inLow, inClose, inVolume []float64, inFastPeriod, inSlowPeriod int) []float64 {
 	outReal := make([]float64, len(inClose))
 
 	if (inFastPeriod < 2) || (inSlowPeriod < 2) {
@@ -175,5 +177,250 @@ func AdOsc(inHigh []float64, inLow []float64, inClose []float64, inVolume []floa
 		outIdx++
 	}
 
+	return outReal
+}
+
+// UltOsc - Ultimate Oscillator
+func UltOsc(inHigh, inLow, inClose []float64, inTimePeriod1 int, inTimePeriod2 int, inTimePeriod3 int) []float64 {
+	outReal := make([]float64, len(inClose))
+
+	usedFlag := make([]int, 3)
+	periods := make([]int, 3)
+	sortedPeriods := make([]int, 3)
+
+	periods[0] = inTimePeriod1
+	periods[1] = inTimePeriod2
+	periods[2] = inTimePeriod3
+
+	for i := 0; i < 3; i++ {
+		longestPeriod := 0
+		longestIndex := 0
+		for j := 0; j < 3; j++ {
+			if (usedFlag[j] == 0) && (periods[j] > longestPeriod) {
+				longestPeriod = periods[j]
+				longestIndex = j
+			}
+		}
+		usedFlag[longestIndex] = 1
+		sortedPeriods[i] = longestPeriod
+	}
+	inTimePeriod1 = sortedPeriods[2]
+	inTimePeriod2 = sortedPeriods[1]
+	inTimePeriod3 = sortedPeriods[0]
+
+	lookbackTotal := 0
+	if inTimePeriod1 > inTimePeriod2 {
+		lookbackTotal = inTimePeriod1
+	}
+	if inTimePeriod3 > lookbackTotal {
+		lookbackTotal = inTimePeriod3
+	}
+	lookbackTotal++
+
+	startIdx := lookbackTotal - 1
+
+	a1Total := 0.0
+	b1Total := 0.0
+	for i := startIdx - inTimePeriod1 + 1; i < startIdx; i++ {
+		tempLT := inLow[i]
+		tempHT := inHigh[i]
+		tempCY := inClose[i-1]
+		trueLow := 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow := inClose[i] - trueLow
+		trueRange := tempHT - tempLT
+		tempDouble := math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a1Total += closeMinusTrueLow
+		b1Total += trueRange
+	}
+
+	a2Total := 0.0
+	b2Total := 0.0
+	for i := startIdx - inTimePeriod2 + 1; i < startIdx; i++ {
+		tempLT := inLow[i]
+		tempHT := inHigh[i]
+		tempCY := inClose[i-1]
+		trueLow := 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow := inClose[i] - trueLow
+		trueRange := tempHT - tempLT
+		tempDouble := math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a2Total += closeMinusTrueLow
+		b2Total += trueRange
+	}
+
+	a3Total := 0.0
+	b3Total := 0.0
+	for i := startIdx - inTimePeriod3 + 1; i < startIdx; i++ {
+		tempLT := inLow[i]
+		tempHT := inHigh[i]
+		tempCY := inClose[i-1]
+		trueLow := 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow := inClose[i] - trueLow
+		trueRange := tempHT - tempLT
+		tempDouble := math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a3Total += closeMinusTrueLow
+		b3Total += trueRange
+	}
+
+	//today := startIdx
+	//outIdx := startIdx
+	//trailingIdx1 := today - inTimePeriod1 + 1
+	//trailingIdx2 := today - inTimePeriod2 + 1
+	//trailingIdx3 := today - inTimePeriod3 + 1
+
+	today := startIdx
+	outIdx := startIdx
+	trailingIdx1 := today - inTimePeriod1 + 1
+	trailingIdx2 := today - inTimePeriod2 + 1
+	trailingIdx3 := today - inTimePeriod3 + 1
+
+	for today < len(inClose) {
+		tempLT := inLow[today]
+		tempHT := inHigh[today]
+		tempCY := inClose[today-1]
+		trueLow := 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow := inClose[today] - trueLow
+		trueRange := tempHT - tempLT
+		tempDouble := math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a1Total += closeMinusTrueLow
+		a2Total += closeMinusTrueLow
+		a3Total += closeMinusTrueLow
+		b1Total += trueRange
+		b2Total += trueRange
+		b3Total += trueRange
+		output := 0.0
+		if !(((-(0.00000000000001)) < b1Total) && (b1Total < (0.00000000000001))) {
+			output += 4.0 * (a1Total / b1Total)
+		}
+		if !(((-(0.00000000000001)) < b2Total) && (b2Total < (0.00000000000001))) {
+			output += 2.0 * (a2Total / b2Total)
+		}
+		if !(((-(0.00000000000001)) < b3Total) && (b3Total < (0.00000000000001))) {
+			output += a3Total / b3Total
+		}
+		tempLT = inLow[trailingIdx1]
+		tempHT = inHigh[trailingIdx1]
+		tempCY = inClose[trailingIdx1-1]
+		trueLow = 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow = inClose[trailingIdx1] - trueLow
+		trueRange = tempHT - tempLT
+		tempDouble = math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a1Total -= closeMinusTrueLow
+		b1Total -= trueRange
+		tempLT = inLow[trailingIdx2]
+		tempHT = inHigh[trailingIdx2]
+		tempCY = inClose[trailingIdx2-1]
+		trueLow = 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow = inClose[trailingIdx2] - trueLow
+		trueRange = tempHT - tempLT
+		tempDouble = math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a2Total -= closeMinusTrueLow
+		b2Total -= trueRange
+		tempLT = inLow[trailingIdx3]
+		tempHT = inHigh[trailingIdx3]
+		tempCY = inClose[trailingIdx3-1]
+		trueLow = 0.0
+		if tempLT < tempCY {
+			trueLow = tempLT
+		} else {
+			trueLow = tempCY
+		}
+		closeMinusTrueLow = inClose[trailingIdx3] - trueLow
+		trueRange = tempHT - tempLT
+		tempDouble = math.Abs(tempCY - tempHT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+		tempDouble = math.Abs(tempCY - tempLT)
+		if tempDouble > trueRange {
+			trueRange = tempDouble
+		}
+
+		a3Total -= closeMinusTrueLow
+		b3Total -= trueRange
+		outReal[outIdx] = 100.0 * (output / 7.0)
+		outIdx++
+		today++
+		trailingIdx1++
+		trailingIdx2++
+		trailingIdx3++
+	}
 	return outReal
 }
