+++ indicators/price.go
@@ -1,0 +1,53 @@
+package indicators
+
+// AvgPrice - Average Price (o+h+l+c)/4
+func AvgPrice(inOpen, inHigh, inLow, inClose []float64) []float64 {
+	outReal := make([]float64, len(inClose))
+	outIdx := 0
+	startIdx := 0
+
+	for i := startIdx; i < len(inClose); i++ {
+		outReal[outIdx] = (inHigh[i] + inLow[i] + inClose[i] + inOpen[i]) / 4
+		outIdx++
+	}
+	return outReal
+}
+
+// MedPrice - Median Price (h+l)/2
+func MedPrice(inHigh, inLow []float64) []float64 {
+	outReal := make([]float64, len(inHigh))
+	outIdx := 0
+	startIdx := 0
+
+	for i := startIdx; i < len(inHigh); i++ {
+		outReal[outIdx] = (inHigh[i] + inLow[i]) / 2.0
+		outIdx++
+	}
+	return outReal
+}
+
+// TypPrice - Typical Price (h+l+c)/3
+func TypPrice(inHigh, inLow, inClose []float64) []float64 {
+	outReal := make([]float64, len(inClose))
+	outIdx := 0
+	startIdx := 0
+
+	for i := startIdx; i < len(inClose); i++ {
+		outReal[outIdx] = (inHigh[i] + inLow[i] + inClose[i]) / 3.0
+		outIdx++
+	}
+	return outReal
+}
+
+// WclPrice - Weighted Close Price
+func WclPrice(inHigh, inLow, inClose []float64) []float64 {
+	outReal := make([]float64, len(inClose))
+	outIdx := 0
+	startIdx := 0
+
+	for i := startIdx; i < len(inClose); i++ {
+		outReal[outIdx] = (inHigh[i] + inLow[i] + (inClose[i] * 2.0)) / 4.0
+		outIdx++
+	}
+	return outReal
+}
